[Modify] NavHiddenMenu.js 가독성을 위한 컴포넌트 분리 진행 - 수직/수평 컨텐츠 부분 분기 - scss 분기 - NavHiddenVertical -> VerticalItems 로 네이밍 변경
